+++ indexer/services/roundtable/src/helpers/adjectives.json
