+++ indexer/services/roundtable/src/tasks/subaccount-username-generator.ts
@@ -1,14 +1,10 @@
-import { randomInt } from 'crypto';
-
 import { logger } from '@dydxprotocol-indexer/base';
 import {
   SubaccountUsernamesTable,
   SubaccountsWithoutUsernamesResult,
 } from '@dydxprotocol-indexer/postgres';
 
-import config from '../config';
-import adjectives from '../lib/adjectives.json';
-import nouns from '../lib/nouns.json';
+import { generateUsername } from '../helpers/usernames-helper';
 
 export default async function runTask(): Promise<void> {
   const subaccounts:
@@ -34,17 +30,4 @@ export default async function runTask(): Promise<void> {
       });
     }
   }
-}
-
-function generateUsername(): string {
-  const randomAdjective: string = adjectives[randomInt(0, adjectives.length)];
-  const randomNoun: string = nouns[randomInt(0, nouns.length)];
-  const randomNumber: string = randomInt(0, 1000).toString().padStart(
-    config.SUBACCOUNT_USERNAME_NUM_RANDOM_DIGITS, '0');
-
-  const capitalizedAdjective: string = randomAdjective.charAt(
-    0).toUpperCase() + randomAdjective.slice(1);
-  const capitalizedNoun: string = randomNoun.charAt(0).toUpperCase() + randomNoun.slice(1);
-
-  return `${capitalizedAdjective}${capitalizedNoun}${randomNumber}`;
 }
